feature: Added data values for XML. You can export node data values to XML. There are restrictions on the types of values that can be exported. All simple types are supported, as well as UA_NodeClass, StatusCode, UA_DateTime, UA_Guid, UA_String (ByteString), UA_NodeId, UA_ExpandedNodeId, UA_QualifiedName, UA_LocalizedText, and UA_DiagnosticInfo. MultidimensionalArrays of all specified types are also supported. Modules have been refactored. Bugs have been fixed and stability has been improved. Added new methods to modules. To use the Open62541 library version <=1.4.5, a compatibility module has been introduced "v_1.4_Compatibility.h". Added new OPCUA client options: max_references_per_node,max_nodes_per_browse, max_browse_continuation_points, max_nodes_per_read. number_of_max_nodes_to_request_data Has been removed.

+++ include/nodesetexporter/common/LoggerBase.h
@@ -72,7 +72,7 @@
 namespace nodesetexporter::common
 {
 
-enum class LogLevel : int
+enum class LogLevel : uint8_t
 {
     All = 0,
     Trace = 1,
@@ -260,7 +260,7 @@ public:
      * @brief Setting a pattern to output data in a certain sequence and with a certain set
      * @param pattern String template type pattern
      */
-    virtual void SetPattern(TString&& /*pattern*/)
+    virtual void SetPattern(TString&& /*pattern*/) // NOLINT(cppcoreguidelines-rvalue-reference-param-not-moved)
     {
         throw std::runtime_error("not resolve");
     };
@@ -291,7 +291,7 @@ public:
     }
 
     explicit LoggerBase(TString&& logger_name)
-        : m_logger_name(std::move(logger_name)){};
+        : m_logger_name(std::move(logger_name)) {};
     virtual ~LoggerBase() = default;
     LoggerBase(LoggerBase const&) noexcept = delete;
     LoggerBase& operator=(LoggerBase const&) = delete;
